eighth commit

+++ src/views/base/ClientInformation.vue
@@ -33,6 +33,26 @@ export default {
       email: null,
       id: null,
       sex: null,
+      id: null,
+      breakfast: null,
+      lunch: null,
+      dinner: null,
+      breakfastMenu: null,
+      lunchMenu: null,
+      dinnerMenu: null,
+      betweenMeals: null,
+      eatOut: null,
+      luxuryFood: null,
+      mainCook: null,
+      allergy: null,
+      likeAndDislike: null,
+      supplement: null,
+      medicine: null,
+      wakeUpTime: null,
+      bedTime: null,
+      timeOfSleeping: null,
+      exerciseHabits: null,
+      activityIntensity: null,
     };
   },
   created() {
@@ -43,24 +63,47 @@ export default {
     this.email = clientData.email;
     this.id = clientData.id;
     this.sex = clientData.sex;
+    this.breakfast = clientData.breakfast;
+    this.lunch = clientData.lunch;
+    this.dinner = clientData.dinner;
+    this.breakfastMenu = clientData.breakfastMenu;
+    this.lunchMenu = clientData.lunchMenu;
+    this.dinnerMenu = clientData.dinnerMenu;
+    this.betweenMeals = clientData.betweenMeals;
+    this.eatOut = clientData.eatOut;
+    this.luxuryFood = clientData.luxuryFood;
+    this.mainCook = clientData.mainCook;
+    this.allergy = clientData.allergy;
+    this.likeAndDislike = clientData.likeAndDislike;
+    this.supplement = clientData.supplement;
+    this.medicine = clientData.medicine;
+    this.wakeUpTime = clientData.wakeUpTime;
+    this.bedTime = clientData.bedTime;
+    this.timeOfSleeping = clientData.timeOfSleeping;
+    this.exerciseHabits = clientData.exerciseHabits;
+    this.activityIntensity = clientData.activityIntensity;
   },
   methods: {
     deleteClient() {
-      firebase
-        .firestore()
-        .collection("client")
-        .doc(this.$store.state.a.userLogin.uid)
-        .collection("client")
-        .doc(this.id)
-        .delete()
-        .then(function () {
-          console.log("Document successfully deleted!");
-        })
-        .catch(function (error) {
-          console.error("Error removing document: ", error);
-        });
+      if (!confirm("この作業は元に戻せません。削除しますか？")) {
+        return false;
+      } else {
+        firebase
+          .firestore()
+          .collection("client")
+          .doc(this.$store.state.a.userLogin.uid)
+          .collection("client")
+          .doc(this.id)
+          .delete()
+          .then(function () {
+            console.log("Document successfully deleted!");
+          })
+          .catch(function (error) {
+            console.error("Error removing document: ", error);
+          });
 
-      this.$router.push("/mypage");
+        this.$router.push("/mypage");
+      }
     },
   },
 };
